# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Pocket Money API
  description: Backend API for managing chores, ledger entries, and settlements in family groups
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.pocket-money.example.com
    description: Production server
...
paths:
...
  /api/v1/groups:
    get:
      tags:
        - Groups
      summary: List user's groups
      description: Returns all groups the authenticated user is a member of
      operationId: listGroups
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Groups
      summary: Create a new group
      description: Creates a new group with the authenticated user as head
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}:
    get:
      tags:
        - Groups
      summary: Get group details
      description: Returns detailed information about a group including members and chores count
      operationId: getGroup
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Group details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}/members:
    get:
      tags:
        - Groups
      summary: List group members
      description: Returns all members of a group
      operationId: listMembers
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}/invite:
    post:
      tags:
        - Groups
      summary: Create group invite
      description: Creates an invite token for joining the group (head only)
      operationId: createInvite
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can create invites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/join:
    post:
      tags:
        - Groups
      summary: Join group
      description: Joins a group using an invite token
      operationId: joinGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRequest'
      responses:
        '200':
          description: Successfully joined group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          description: Invalid token, token expired, or already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          description: Group name
      required:
        - name
    GroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Group ID
        name:
          type: string
          description: Group name
        head_user_id:
          type: string
          format: uuid
          description: ID of the group head (admin)
        created_at:
          type: string
          format: date-time
          description: Group creation timestamp
      required:
        - id
        - name
        - head_user_id
        - created_at
    GroupDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Group ID
        name:
          type: string
          description: Group name
        head_user_id:
          type: string
          format: uuid
          description: ID of the group head (admin)
        created_at:
          type: string
          format: date-time
          description: Group creation timestamp
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberResponse'
          description: List of group members
        chores_count:
          type: integer
          description: Number of chores in the group
      required:
        - id
        - name
        - head_user_id
        - created_at
        - members
        - chores_count
    MemberResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: User ID
        name:
          type: string
          description: User's full name
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [head, member]
          description: User's role in the group
        joined_at:
          type: string
          format: date-time
          description: Timestamp when user joined the group
      required:
        - user_id
        - name
        - email
        - role
        - joined_at
    InviteRequest:
      type: object
      properties:
        expires_in_days:
          type: integer
          minimum: 1
          default: 7
          description: Number of days until invite expires
    InviteResponse:
      type: object
      properties:
        invite_url:
          type: string
          description: Full invite URL
        token:
          type: string
          description: Invite token
        expires_at:
          type: string
          format: date-time
          description: Expiration timestamp
      required:
        - invite_url
        - token
        - expires_at
    JoinRequest:
      type: object
      properties:
        token:
          type: string
          description: Invite token
      required:
        - token
...